% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-construction.R
\name{as_ggside}
\alias{as_ggside}
\title{Explicit conversion to ggside object}
\usage{
as_ggside(x, ...)
}
\arguments{
\item{x}{an object to convert}

\item{...}{unused argument}

\item{ggside}{new ggside object to add}
}
\description{
Function is only exported for possible extensions to ggside. ggplot2 objects
are implicitly converted to ggside objects by 'adding' a ggside object
such as a \code{ggside_layer} object.
}
