% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.r
\name{class_ggside}
\alias{class_ggside}
\title{Class Definitions}
\description{
Class Definitions
}
\section{S7 classes}{

\itemize{
\item \code{class_ggside} is a subclass of
\link[ggplot2:ggplot2]{foo} and
is used to represent a ggplot object with ggside options.
}
}

