% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.r
\docType{data}
\name{class_ggside_opt}
\alias{class_ggside_opt}
\title{Class Definitions}
\description{
This documentation provides an overview of the S7 and ggproto classes used in the \code{ggside} package.
}
\section{ggproto classes}{

\itemize{
\item \code{class_ggside_opt} is a subclass of \code{class_ggproto} and is
more described in the \link[ggside:ggside]{ggside-options} documentation.
}
}

\keyword{datasets}
