% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.r
\docType{data}
\name{class_ggside_scale}
\alias{class_ggside_scale}
\title{Class Definitions}
\description{
Class Definitions
}
\section{ggproto classes}{

\itemize{
\item \code{class_ggside_scale} is a subclass of \code{class_ggproto} and is
more described in the \link[ggside:ggside_scales]{ggside-scales}
documentation.
}
}

\keyword{datasets}
