% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-construction.R
\name{ggplot_add.ggside_layer}
\alias{ggplot_add.ggside_layer}
\title{Adding ggside layers to ggplot}
\usage{
\method{ggplot_add}{ggside_layer}(object, plot, object_name)
}
\value{
A modified ggplot object
}
\description{
Adding ggside layers to ggplot
}
\seealso{
\link[ggplot]{ggplot_add}
}
