% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sideCoord-Cartesian.R
\name{as_ggsideCoord}
\alias{as_ggsideCoord}
\alias{as_ggsideCoord.default}
\alias{as_ggsideCoord.CoordCartesian}
\alias{as_ggsideCoord.CoordSide}
\alias{as_ggsideCoord.CoordTrans}
\alias{as_ggsideCoord.CoordFixed}
\title{Coord Compatible with ggside}
\usage{
as_ggsideCoord(coord)

\method{as_ggsideCoord}{default}(coord)

\method{as_ggsideCoord}{CoordCartesian}(coord)

\method{as_ggsideCoord}{CoordSide}(coord)

\method{as_ggsideCoord}{CoordTrans}(coord)

\method{as_ggsideCoord}{CoordFixed}(coord)
}
\arguments{
\item{coord}{coord ggproto Object to replace}
}
\description{
S3 class that converts old Coord into one that
is compatible with ggside. Can also update
ggside on the object. Typically, the new ggproto
will inherit from the object being replaced.
}
