% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-other_utils.r, R/geom-sidebar.r,
%   R/geom-sideboxplot.r, R/geom-sidecol.r, R/geom-sidedensity.r,
%   R/geom-sidefunction.r, R/geom-sidelabel.r, R/geom-sideline.r, R/geom-sidepath.r,
%   R/geom-sidepoint.r, R/geom-sidesegment.r, R/geom-sidetext.r,
%   R/geom-sidetile.r, R/geom-sideviolin.r, R/ggside-ggproto.r
\docType{data}
\name{use_xside_aes}
\alias{use_xside_aes}
\alias{use_yside_aes}
\alias{parse_side_aes}
\alias{GeomXsidebar}
\alias{GeomYsidebar}
\alias{GeomXsideboxplot}
\alias{GeomYsideboxplot}
\alias{GeomXsidecol}
\alias{GeomYsidecol}
\alias{GeomXsidedensity}
\alias{GeomYsidedensity}
\alias{GeomXsidelabel}
\alias{GeomYsidelabel}
\alias{GeomXsidefunction}
\alias{GeomYsidefunction}
\alias{GeomXsideline}
\alias{GeomYsideline}
\alias{GeomXsidepath}
\alias{GeomYsidepath}
\alias{GeomXsidepoint}
\alias{GeomYsidepoint}
\alias{GeomXsidesegment}
\alias{GeomYsidesegment}
\alias{GeomXsidetext}
\alias{GeomYsidetext}
\alias{GeomXsidetile}
\alias{GeomYsidetile}
\alias{GeomXsideviolin}
\alias{GeomYsideviolin}
\alias{ggside-ggproto-geoms}
\title{Extending base ggproto classes for ggside}
\usage{
use_xside_aes(data)

use_yside_aes(data)

parse_side_aes(data, params)
}
\arguments{
\item{data}{data passed internally}

\item{params}{params available to ggproto object}
}
\value{
ggproto object that is usually passed to \link[ggplot2]{layer}
}
\description{
These ggproto classes are slightly modified from their
respective inherited \link[ggplot2]{ggproto} class. The
biggest difference is exposing 'x/yfill', 'x/ycolour', and
'x/ycolor' as viable aesthetic mappings.
}
\keyword{datasets}
