% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-other_utils.r, R/ggside-ggproto.r
\name{use_xside_aes}
\alias{use_xside_aes}
\alias{use_yside_aes}
\alias{parse_side_aes}
\alias{ggside-ggproto}
\title{Extending base ggproto classes for ggside}
\usage{
use_xside_aes(data)

use_yside_aes(data)

parse_side_aes(data, params)
}
\arguments{
\item{data}{data passed internally}

\item{params}{params available to ggproto object}
}
\description{
TBD
}
