% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggside-package.r
\docType{package}
\name{ggside-package}
\alias{ggside-package}
\title{ggside: Side Grammar Graphics}
\description{
The grammar of graphics as shown in 'ggplot2' has provided an expressive API for users to build plots. 'ggside' extends 'ggplot2' by allowing users to add graphical information about one of the main panel's axis using a familiar 'ggplot2' style API with tidy data. This package is particularly useful for visualizing metadata on a discrete axis, or summary graphics on a continuous axis such as a boxplot or a density distribution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jtlandis/ggside}
  \item Report bugs at \url{https://github.com/jtlandis/ggside/issues}
}

}
\author{
\strong{Maintainer}: Justin Landis \email{jtlandis314@gmail.com} (\href{https://orcid.org/0000-0001-5501-4934}{ORCID})

}
\keyword{internal}
