% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-sides-.R
\name{ggside-scales}
\alias{ggside-scales}
\alias{ggside_scales}
\title{Specifying side scales}
\description{
The \link{xside} and \link{yside} variants of \link[ggplot2:Geom]{geoms} are plotted
along the x-axis and y-axis respectively of their main panel's data mapping.
The positional scale here is shared between the main panel and the side
panel. The related positional scale type of the side panel, i.e. the y axis
of the xside panel (xsidey) or the x axis of the yside panel (ysidex), is
determined automatically by \code{ggplot2} default scales. However, you can
override this by using the \link[=ggside-scales-continuous]{continuous} or
\link[=ggside-scales-discrete]{discrete} variants within \code{ggside}. This
allows the user to select the scale type or transform most appropriate for
their side panels.
}
