% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ggside-theme}
\alias{ggside-theme}
\title{ggside custom themes}
\description{
the following are the theme elements defined in
\link{ggside}.
}
\section{Elements}{

\itemize{
\item ggside.panel.scale - sets the scaling of side panels relative
to the plotting width of the main panels. Default is set to 0.1, i.e. 0.1/1
\item ggside.panel.scale.x - same as ggside.panel.scale except only
for the xside panel.
\item ggside.panel.scale.y - same as ggside.panel.scale except only
for the yside panel.
\item ggside.panel.spacing - sets how much spacing should be used between
the main panels and the side panels. default is unit(2,"pt")
\item ggside.panel.spacing.x - same as ggside.panel.spacing except only
for the space between the main panel and the yside panel.
\item ggside.panel.spacing.y - same as ggside.panel.spacing except only
for the space between the main panel and the xside panel.
}
}

