% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggside.R
\docType{package}
\name{ggside}
\alias{ggside}
\title{ggside options}
\usage{
ggside(x.pos = "top", y.pos = "right", scales = "fixed", collapse = NULL)
}
\arguments{
\item{x.pos}{x side panel can either take "top" or "bottom"}

\item{y.pos}{y side panel can either take "right" or "left"}

\item{scales}{Determines side panel's unaligned axis
scale. Inputs are similar to facet_* scales function. Default
is set to "fixed", but "free_x", "free_y" and "free" are
acceptable inputs. For example, xside panels are aligned to
the x axis of the main panel. Setting "free" or "free_y" will
cause all y scales of the x side Panels to be independent.}

\item{collapse}{Determins if side panels should be collapsed into
a single panel. Set "x" to collapse all x side panels, set "y" to
collapse all y side panels, set "all" to collapse both x and y
side panels.}
}
\value{
a object of class 'ggside_options' or to be added to a ggpot
}
\description{
Set characteristics of side panels
}
\seealso{
For more information regarding the ggside api: see \link{xside} or \link{yside}
}
