% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor-2.R
\name{ggside_geom}
\alias{ggside_geom}
\title{ggside geom constructor}
\usage{
ggside_geom(class_name = NULL, geom = NULL, side = NULL, env = parent.frame())
}
\arguments{
\item{class_name}{New class name for the ggproto object}

\item{geom}{The Geom ggproto to inherit from}

\item{side}{should the resulting object be configured for x or y}
}
\description{
utility function to make a ggside Geom
}
