% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side-layout-.r
\name{ggside_layout}
\alias{ggside_layout}
\title{Construct ggside layout}
\usage{
ggside_layout(layout)
}
\arguments{
\item{layout}{a ggproto Layout object}
}
\description{
Creates a new layout object required for ggside functionality
}
