% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor-.R
\name{ggside_scales}
\alias{ggside_scales}
\title{Creates new ScaleList
Modifies the \verb{$add} method such that
when a positional scale is added it is
recast with \code{new_pos_scale}}
\usage{
ggside_scales(scales, ggside)
}
\description{
Creates new ScaleList
Modifies the \verb{$add} method such that
when a positional scale is added it is
recast with \code{new_pos_scale}
}
