% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggside.R
\name{is.ggside}
\alias{is.ggside}
\alias{is.ggside_layer}
\alias{is.ggside_options}
\alias{is.ggside_scale}
\title{Check ggside objects}
\usage{
is.ggside(x)

is.ggside_layer(x)

is.ggside_options(x)

is.ggside_scale(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
A logical value
}
\description{
Check ggside objects
}
