% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-utilities-.R
\name{local_union_scale_aes}
\alias{local_union_scale_aes}
\title{Temporarily changes the first element of a list to contain the
union aesthetics of all other elements}
\usage{
local_union_scale_aes(scales, frame = parent.frame())
}
\arguments{
\item{scales}{a list containing a ggproto Scale object}
}
\description{
Temporarily changes the first element of a list to contain the
union aesthetics of all other elements
}
