% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-utilities-.R
\name{local_vanilla_scale_aes}
\alias{local_vanilla_scale_aes}
\title{Temporarily changes any \code{xside} or \code{yside} prefix to be removed
from scale$aesthetic. Values are returned when exiting the frame this
function was called in.}
\usage{
local_vanilla_scale_aes(scale, frame = parent.frame())
}
\arguments{
\item{scale}{a ggproto Scale object}
}
\description{
Temporarily changes any \code{xside} or \code{yside} prefix to be removed
from scale$aesthetic. Values are returned when exiting the frame this
function was called in.
}
