% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-utilities-.R
\name{new_default_aes}
\alias{new_default_aes}
\title{utility to recode default aesthetics of a geom.}
\usage{
new_default_aes(geom, side)
}
\value{
an object of class 'uneval'
}
\description{
utility to recode default aesthetics of a geom.
}
