% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor-.R
\name{new_pos_scale}
\alias{new_pos_scale}
\title{positional scales in \code{ggside} will contain
the \code{xsidey} and \code{ysidex} in their aesthetics
to ensure all data are trained properly.
To ensure each scale retains its own transform,
these scales will ignore \code{xsidey} and \code{ysidex}
aesthetics when \verb{$transform_df} is called.}
\usage{
new_pos_scale(scale)
}
\description{
positional scales in \code{ggside} will contain
the \code{xsidey} and \code{ysidex} in their aesthetics
to ensure all data are trained properly.
To ensure each scale retains its own transform,
these scales will ignore \code{xsidey} and \code{ysidex}
aesthetics when \verb{$transform_df} is called.
}
