% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_rescale.r
\docType{data}
\name{position_range}
\alias{position_range}
\alias{PositionRescale}
\title{Rescale x or y onto new range in margin}
\format{
An object of class \code{PositionRescale} (inherits from \code{Position}, \code{ggproto}, \code{gg}) of length 10.
}
\description{
Rescale x or y onto new range in margin
}
\keyword{datasets}
