% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-utilities-.R
\name{pull_side}
\alias{pull_side}
\title{utility to pull out an aesthetic we care about.
helps code around the \code{|} aesthetics}
\usage{
pull_side(x, i)
}
\value{
a character vector
}
\description{
utility to pull out an aesthetic we care about.
helps code around the \code{|} aesthetics
}
