% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-xycolour.R
\name{scale_xcolour}
\alias{scale_xcolour}
\alias{scale_ycolour}
\alias{scale_xcolor}
\alias{scale_ycolor}
\alias{scale_xcolour_hue}
\alias{scale_xcolour_manual}
\alias{scale_xcolor_manual}
\alias{scale_xcolour_gradient}
\alias{scale_xcolor_gradientn}
\alias{scale_xcolour_gradientn}
\alias{scale_xcolour_discrete}
\alias{scale_xcolor_discrete}
\alias{scale_xcolour_continuous}
\alias{scale_xcolor_continuous}
\title{Scales for the *colour aesthetics}
\value{
returns a ggproto object to be added to a ggplot
}
\description{
These are the various scales that can be applied to the xsidebar or ysidebar
colour aesthetics, such as xcolour and ycolour. They have the same usage as
existing standard ggplot2 scales.
}
\section{Related Functions}{


\itemize{
\item scale_xcolour_hue
\item scale_ycolour_hue
\item scale_xcolour_discrete
\item scale_ycolour_discrete
\item scale_xcolour_continuous
\item scale_ycolour_continuous
\item scale_xcolour_manual
\item scale_ycolour_manual
\item scale_xcolour_gradient
\item scale_ycolour_gradient
\item scale_xcolour_gradientn
\item scale_ycolour_gradientn
}
}

