% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-xyfill.R
\name{scale_xfill}
\alias{scale_xfill}
\alias{scale_yfill}
\alias{scale_xfill_hue}
\alias{scale_xfill_manual}
\alias{scale_xfill_gradient}
\alias{scale_xfill_gradientn}
\alias{scale_xfill_discrete}
\alias{scale_xfill_continuous}
\alias{scale_yfill_gradientn}
\title{Scales for the *fill aesthetics}
\value{
returns a ggproto object to be added to a ggplot
}
\description{
These are the various scales that can be applied to the xsidebar or ysidebar
fill aesthetics, such as xfill and yfill. They have the same usage as
existing standard ggplot2 scales.
}
\section{Related Functions}{


\itemize{
\item scale_xfill_hue
\item scale_yfill_hue
\item scale_xfill_discrete
\item scale_yfill_discrete
\item scale_xfill_continuous
\item scale_yfill_continuous
\item scale_xfill_manual
\item scale_yfill_manual
\item scale_xfill_gradient
\item scale_yfill_gradient
\item scale_xfill_gradientn
\item scale_yfill_gradientn
}
}

