% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-xycolour.R
\name{scale_ycolour_hue}
\alias{scale_ycolour_hue}
\alias{scale_ycolour_manual}
\alias{scale_ycolor_manual}
\alias{scale_ycolour_gradient}
\alias{scale_ycolour_gradientn}
\alias{scale_ycolor_gradientn}
\alias{scale_ycolour_discrete}
\alias{scale_ycolor_discrete}
\alias{scale_ycolour_continuous}
\alias{scale_ycolor_continuous}
\title{scale_ycolour_hue}
\description{
scale_ycolour_hue

scale_ycolour_manual

scale_ycolour_gradient

scale_ycolour_discrete

scale_ycolour_discrete

scale_ycolour_continuous

scale_ycolour_continuous
}
