% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_yfill_hue}
\alias{scale_yfill_hue}
\alias{scale_yfill_gradient}
\alias{scale_yfill_discrete}
\alias{scale_yfill_continuous}
\title{scale_yfill_hue}
\description{
scale_yfill_hue

scale_yfill_gradient

scale_yfill_discrete

scale_yfill_continuous
}
