% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.r
\name{standardise_aes_names}
\alias{standardise_aes_names}
\title{This function standardises aesthetic names by converting `color` to `colour`
(also in substrings, e.g. `point_color` to `point_colour`) and translating old style
R names to ggplot names (eg. `pch` to `shape`, `cex` to `size`).}
\usage{
standardise_aes_names(x)
}
\arguments{
\item{x}{Character vector of aesthetics names, such as `c("colour", "size", "shape")`.}
}
\value{
Character vector of standardised names.
}
\description{
This function standardises aesthetic names by converting `color` to `colour`
(also in substrings, e.g. `point_color` to `point_colour`) and translating old style
R names to ggplot names (eg. `pch` to `shape`, `cex` to `size`).
}
\keyword{internal}
