% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-utilities-.R
\name{standardise_panel_params}
\alias{standardise_panel_params}
\title{Standardize the panel params such that any \code{xside} or \code{yside} prefixes
are removed from the object}
\usage{
standardise_panel_params(built)
}
\arguments{
\item{panel_params}{the panel_params element of the Layout ggproto object}

\item{layout}{the layout element of the Layout ggproto object}
}
\description{
Standardize the panel params such that any \code{xside} or \code{yside} prefixes
are removed from the object
}
