% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\docType{data}
\name{stat_summarise}
\alias{stat_summarise}
\alias{stat_summarize}
\alias{StatSummarise}
\alias{StatSummarize}
\title{Summarise by grouping variable}
\format{
An object of class \code{Summarise} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 5.

An object of class \code{Summarize} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
stat_summarise(
  mapping = NULL,
  data = NULL,
  geom = "bar",
  position = "identity",
  ...,
  fun = NULL,
  fun.args = list(),
  show.legend = NA,
  inherit.aes = TRUE
)

stat_summarize(
  mapping = NULL,
  data = NULL,
  geom = "bar",
  position = "identity",
  ...,
  fun = NULL,
  fun.args = list(),
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
Applies a function to a specified grouping variable
}
\examples{
#library(tidyr)
i <- gather(iris,"key","value",-Species)
ggplot(i, aes(Species, fill = key, domain = value)) +
   geom_bar(aes(y = after_stat(summarise)), stat = "summarise") +
   stat_summarise(aes(y = after_stat(summarise),
                  label = after_stat(summarise)),
                  position = position_stack(vjust = .5), geom = "text")
ggplot(i, aes(key, fill = Species, domain = value)) +
   geom_bar(aes(y = after_stat(summarise)), stat = "summarise") +
   stat_summarise(aes(y = after_stat(summarise),
                  label = after_stat(summarise)),
                  position = position_stack(vjust = .5), geom = "text")
}
\keyword{datasets}
