% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.r
\docType{data}
\name{geom_xsidebar}
\alias{geom_xsidebar}
\alias{GeomXSideBar}
\alias{geom_ysidebar}
\alias{GeomYSideBar}
\title{geom_*sidebar: Plot a sidebar along an axis}
\usage{
geom_xsidebar(
  mapping = NULL,
  data = NULL,
  na.rm = FALSE,
  show.legend = NA,
  position = "identity",
  stat = "sidebar",
  inherit.aes = TRUE,
  ...
)

geom_ysidebar(
  mapping = NULL,
  data = NULL,
  na.rm = FALSE,
  show.legend = NA,
  position = "identity",
  stat = "sidebar",
  inherit.aes = TRUE,
  ...
)

GeomYSideBar
}
\arguments{
\item{mapping}{Set of aesthetic mappings reated by [`aes()`] or
[`aes_()`]. If specified and `inherit.aes = TRUE`, it will be
combined with the default mapping at the top level of the plot.
`mapping` must be supplied in this layer if ther is no plot mapping.}

\item{data}{The data to be displayed in this layer. If `NULL`, the default,
then the data is inherited from the plot data as specified i nthe call to [`ggplot()`].}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{inherit.aes}{If `FALSE`, overrides the default aesthetics,
rather than combining with them.}

\item{...}{other arguments passed on to [`layer()`]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`color = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}
}
\description{
geom_*sidebar: Plot a sidebar along an axis
}
\section{Aesthetics}{


Required aesthetics are in bold.

\itemize{
\item \strong{x}
\item \strong{y}
\item \emph{xfill} Fill color of the xsidebar
\item \emph{yfill} Fill color of the ysidebar
\item \emph{width} specifies the width of each bar
\item \emph{height} specifies the height of each bar
\item \emph{alpha} Transparency level of `xfill` or `yfill` depending on which [`geom_*sidbar`]
\item \emph{size} size of the border line. --uneeded??
\item \emph{location} Specifies where the sidebar should be placed.
geom_xsidebar may specify either "bottom" or "top" and
geom_ysidebar may specify either "left" or "right.
"bottom" and "left" are defaults of geom_xsidebar and
geom_ysidebar respecitively.

}
}

\keyword{datasets}
