% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-depricated.R
\name{ggside-deprecated}
\alias{ggside-deprecated}
\alias{as_ggsideFacet}
\alias{as_ggsideCoord}
\title{Deprecated Functions}
\description{
The following functions have been deprecated.

as_ggsideFacet <- \link{ggside_facet}
as_ggsideCoord <- \link{ggside_coord}
}
