% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-sides-.R
\name{ggside-scales-discrete}
\alias{ggside-scales-discrete}
\alias{scale_xsidey_discrete}
\alias{scale_ysidex_discrete}
\title{Position scales for discrete data ggside scales}
\arguments{
\item{expand}{For position scales, a vector of range expansion constants used to add some
padding around the data to ensure that they are placed some distance
away from the axes. Use the convenience function \code{\link[ggplot2:expansion]{expansion()}}
to generate the values for the \code{expand} argument. The defaults are to
expand the scale by 5\% on each side for continuous variables, and by
0.6 units on each side for discrete variables.}

\item{guide}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}

\item{position}{For position scales, The position of the axis.
\code{left} or \code{right} for y axes, \code{top} or \code{bottom} for x axes.}
}
\value{
ggside_scale object inheriting from ggplot2::ScaleDiscretePosition
}
\description{
The \link{xside} and \link{yside} variants of \link[ggplot2]{scale_x_discrete}/\link[ggplot2]{scale_y_discrete}.
\link{scale_xsidey_discrete} enables better control on how the y-axis is rendered on the xside panel and
\link{scale_ysidex_discrete} enables better control on how the x-axis is rendered on the yside panel.
}
